refactor: 컨트롤러 필드 컨벤션 적용
