+++ src/App.tsx
@@ -1,7 +1,7 @@
 import Everytime from "@/components/pages/Everytime";
 import LoginPage from "@/components/pages/LoginPage";
 import Payment from "@/components/pages/Payment";
-import PersonalInfo from "@/components/pages/PersonalInfo";
+import PersonalInfo from "@/components/pages/PersonalInfo/PersonalInfo";
 import Survey from "@/components/pages/Survey";
 import { Progress } from "@/components/ui/progress";
 import { useState } from "react";
