+++ src/components/pages/PersonalInfo/field/studentGrade.tsx
@@ -7,8 +7,7 @@ import {
 } from "@/components/ui/select";
 import { Label } from "@/components/ui/label";
 import { forwardRef } from "react";
-import type { Grade } from "@/types/api/member";
-import { useFormContext } from "react-hook-form";
+import { useControllerField } from "../PersonalInfo.ControlledField";
 
 //학년 필드 배열
 const grades = [
@@ -19,21 +18,18 @@ const grades = [
   { value: "FIVE", label: "5학년" },
 ];
 
-interface StudentGradeProps extends React.ComponentPropsWithoutRef<typeof Select>{
-  error?: string;
-  value?: Grade;
-  onChange?: (value: Grade) => void;
+interface StudentGradeProps {
+  name: string; // name prop 추가
 }
 
 export const StudentGrade = forwardRef<HTMLDivElement, StudentGradeProps>(
-  ({ error, value, onChange, ...props }, ref)=>{
-    const { formState } = useFormContext();
-    const isValid = !formState.errors.grade;
+  ({ name, ...props }, ref)=>{
+    const { field, error, isValid } = useControllerField({ name });
 
   return(
     <div className="space-y-2" {...props} ref={ref}>
         <Label htmlFor="grade">모집 학기 기준 학년</Label>
-        <Select value={value} onValueChange={onChange}>
+        <Select value={field.value} onValueChange={field.onChange}>
           <SelectTrigger className={error && !isValid ? "border-red-500" : ""}>
             <SelectValue placeholder="학년 선택" />
           </SelectTrigger>
@@ -51,3 +47,5 @@ export const StudentGrade = forwardRef<HTMLDivElement, StudentGradeProps>(
       </div>
   );
 });
+
+StudentGrade.displayName = "StudentGrade";
